requirements.txt
numpy==1.16.1
opencv-python==4.0.0.21
Pillow==5.4.1
pytesseract==0.2.6
six==1.12.0
torch==1.0.1.post2
torchvision==0.2.2.post3
